+++ src/Explorer/Notebook/NotebookClientV2.ts
@@ -34,8 +34,8 @@ import * as TelemetryProcessor from "../../Shared/Telemetry/TelemetryProcessor";
 import { userContext } from "../../UserContext";
 import configureStore from "./NotebookComponent/store";
 import { CdbAppState, makeCdbRecord } from "./NotebookComponent/types";
-import HTML from "./NotebookRenderer/outputs/html";
-import JavaScript from "./NotebookRenderer/outputs/javascript";
+import SandboxJavaScript from "./NotebookRenderer/outputs/SandboxJavaScript";
+import SanitizedHTML from "./NotebookRenderer/outputs/SanitizedHTML";
 
 export type KernelSpecsDisplay = { name: string; displayName: string };
 
@@ -168,8 +168,10 @@ export class NotebookClientV2 {
               "application/vnd.vega.v5+json": NullTransform,
               "application/vdom.v1+json": TransformVDOM,
               "application/json": Media.Json,
-              "application/javascript": userContext.features.sandboxNotebookOutputs ? JavaScript : Media.JavaScript,
-              "text/html": userContext.features.sandboxNotebookOutputs ? HTML : Media.HTML,
+              "application/javascript": userContext.features.sandboxNotebookOutputs
+                ? SandboxJavaScript
+                : Media.JavaScript,
+              "text/html": userContext.features.sandboxNotebookOutputs ? SanitizedHTML : Media.HTML,
               "text/markdown": Media.Markdown,
               "text/latex": Media.LaTeX,
               "image/svg+xml": Media.SVG,
